+++ src/ent/display_states.hpp
@@ -3,7 +3,7 @@
 #include "ent/display_support.hpp"
 #include "aframe_id.hpp"
 
-#include <cstddef>
+#include <cstdint>
 #include <optional>
 
 namespace splonks {
@@ -12,7 +12,7 @@ struct DisplayStateAFrameSelection {
     AFrameId anim_id = kInvalidAFrameId;
     bool animate = true;
     bool has_forced_frame = false;
-    std::size_t forced_frame = 0;
+    std::uint32_t forced_frame = 0;
 };
 
 std::optional<DisplayStateAFrameSelection> GetAFrameSelectionForDisplayState(
